load hdfs validation tool from registry
