+++ integration/tools/validation/pom.xml
@@ -13,9 +13,9 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>org.alluxio</groupId>
-        <artifactId>alluxio-integration-tools</artifactId>
-        <version>2.3.0-SNAPSHOT</version>
+      <groupId>org.alluxio</groupId>
+      <artifactId>alluxio-integration-tools</artifactId>
+      <version>2.3.0-SNAPSHOT</version>
     </parent>
     <artifactId>alluxio-integration-tools-validation</artifactId>
     <packaging>jar</packaging>
@@ -23,45 +23,82 @@
     <description>Tools to validation the underlying environment for Alluxio</description>
 
     <properties>
-        <!-- The following paths need to be defined here as well as in the parent pom so that mvn can -->
-        <!-- run properly from sub-project directories -->
-        <build.path>${project.parent.parent.parent.basedir}/build</build.path>
-        <failIfNoTests>false</failIfNoTests>
+      <!-- The following paths need to be defined here as well as in the parent pom so that mvn can -->
+      <!-- run properly from sub-project directories -->
+      <build.path>${project.parent.parent.parent.basedir}/build</build.path>
+      <failIfNoTests>false</failIfNoTests>
+      <lib.jar.name>${project.artifactId}-${project.version}.jar</lib.jar.name>
     </properties>
 
     <dependencies>
-        <!-- External dependencies -->
-        <dependency>
-            <groupId>com.google.guava</groupId>
-            <artifactId>guava</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>jline</groupId>
-            <artifactId>jline</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.beust</groupId>
-            <artifactId>jcommander</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.commons</groupId>
-            <artifactId>commons-compress</artifactId>
-        </dependency>
+      <!-- External dependencies -->
+      <dependency>
+        <groupId>com.google.guava</groupId>
+        <artifactId>guava</artifactId>
+      </dependency>
+      <dependency>
+        <groupId>jline</groupId>
+        <artifactId>jline</artifactId>
+      </dependency>
+      <dependency>
+        <groupId>com.beust</groupId>
+        <artifactId>jcommander</artifactId>
+      </dependency>
+      <dependency>
+        <groupId>org.apache.commons</groupId>
+        <artifactId>commons-compress</artifactId>
+      </dependency>
 
-        <!-- Internal dependencies -->
-        <dependency>
-            <groupId>org.alluxio</groupId>
-            <artifactId>alluxio-core-common</artifactId>
-            <version>${project.version}</version>
-        </dependency>
+      <!-- Internal dependencies -->
+      <dependency>
+        <groupId>org.alluxio</groupId>
+        <artifactId>alluxio-core-common</artifactId>
+        <version>${project.version}</version>
+      </dependency>
 
-        <!-- Test dependencies -->
-        <dependency>
-            <groupId>org.alluxio</groupId>
-            <artifactId>alluxio-core-common</artifactId>
-            <version>${project.version}</version>
-            <type>test-jar</type>
-            <scope>test</scope>
-        </dependency>
+      <!-- Test dependencies -->
+      <dependency>
+        <groupId>org.alluxio</groupId>
+        <artifactId>alluxio-core-common</artifactId>
+        <version>${project.version}</version>
+        <type>test-jar</type>
+        <scope>test</scope>
+      </dependency>
     </dependencies>
+
+      <build>
+    <plugins>
+      <plugin>
+        <groupId>com.coderplus.maven.plugins</groupId>
+        <artifactId>copy-rename-maven-plugin</artifactId>
+        <executions>
+          <execution>
+            <id>copy-and-rename-file</id>
+            <phase>install</phase>
+            <goals>
+              <goal>copy</goal>
+            </goals>
+            <configuration>
+              <sourceFile>${basedir}/target/${project.artifactId}-${project.version}.jar</sourceFile>
+              <destinationFile>${project.parent.parent.parent.basedir}/lib/${lib.jar.name}</destinationFile>
+            </configuration>
+          </execution>
+        </executions>
+      </plugin>
+      <plugin>
+        <artifactId>maven-clean-plugin</artifactId>
+        <configuration>
+          <filesets>
+            <fileset>
+              <directory>${basedir}/../../../lib</directory>
+              <includes>
+                <include>**/${project.artifactId}-*.jar</include>
+              </includes>
+            </fileset>
+          </filesets>
+        </configuration>
+      </plugin>
+    </plugins>
+  </build>
+
 </project>
