+++ integration/tools/validation/src/main/java/alluxio/cli/RamDiskMountPrivilegeValidationTask.java
@@ -25,7 +25,7 @@ import java.util.Map;
 /**
  * Task for validating whether worker has enough privilege to mount RAM disk.
  */
-public final class RamDiskMountPrivilegeValidationTask implements ValidationTask {
+public final class RamDiskMountPrivilegeValidationTask extends AbstractValidationTask {
   private final AlluxioConfiguration mConf;
 
   /**
