+++ integration/tools/validation/src/main/java/alluxio/cli/SshValidationTask.java
@@ -23,7 +23,7 @@ import java.util.Set;
 /**
  * Task for validating SSH reachability.
  */
-public final class SshValidationTask implements ValidationTask {
+public final class SshValidationTask extends AbstractValidationTask {
   private final AlluxioConfiguration mConf;
 
     /**
