+++ integration/tools/validation/src/main/java/alluxio/cli/ValidateEnv.java
@@ -12,24 +12,38 @@
 package alluxio.cli;
 
 import alluxio.Constants;
+import alluxio.cli.ClusterConfConsistencyValidationTask;
 import alluxio.cli.hdfs.HdfsConfParityValidationTask;
 import alluxio.cli.hdfs.HdfsConfValidationTask;
 import alluxio.cli.hdfs.HdfsProxyUserValidationTask;
 import alluxio.cli.hdfs.HdfsVersionValidationTask;
+import alluxio.cli.NativeLibValidationTask;
+import alluxio.cli.PortAvailabilityValidationTask;
+import alluxio.cli.RamDiskMountPrivilegeValidationTask;
 import alluxio.cli.hdfs.SecureHdfsValidationTask;
+import alluxio.cli.SshValidationTask;
+import alluxio.cli.StorageSpaceValidationTask;
+import alluxio.cli.UfsDirectoryValidationTask;
+import alluxio.cli.UfsSuperUserValidationTask;
+import alluxio.cli.UserLimitValidationTask;
+import alluxio.cli.ValidationTask;
 import alluxio.conf.AlluxioConfiguration;
+import alluxio.conf.InstancedConfiguration;
 import alluxio.conf.PropertyKey;
+import alluxio.exception.status.InvalidArgumentException;
 import alluxio.util.CommonUtils;
 import alluxio.util.ConfigurationUtils;
 import alluxio.util.network.NetworkAddressUtils.ServiceType;
 
 import org.apache.commons.cli.CommandLine;
+import org.apache.commons.cli.CommandLineParser;
+import org.apache.commons.cli.DefaultParser;
+import org.apache.commons.cli.HelpFormatter;
 import org.apache.commons.cli.Option;
 import org.apache.commons.cli.Options;
-import org.reflections.Reflections;
+import org.apache.commons.cli.ParseException;
 
 import java.io.IOException;
-import java.lang.reflect.Field;
 import java.util.Arrays;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -43,10 +57,31 @@ import java.util.Map;
  * Utility for checking Alluxio environment.
  */
 // TODO(yanqin): decouple ValidationTask implementations for easier dependency management
-public final class EnvValidationTool {
+public final class ValidateEnv {
+  private static final String USAGE = "validateEnv COMMAND [NAME] [OPTIONS]\n\n"
+          + "Validate environment for Alluxio.\n\n"
+          + "COMMAND can be one of the following values:\n"
+          + "local:   run all validation tasks on local\n"
+          + "master:  run master validation tasks on local\n"
+          + "worker:  run worker validation tasks on local\n"
+          + "all:     run corresponding validation tasks on all master nodes and worker nodes\n"
+          + "masters: run master validation tasks on all master nodes\n"
+          + "workers: run worker validation tasks on all worker nodes\n\n"
+          + "list:    list all validation tasks\n\n"
+          + "For all commands except list:\n"
+          + "NAME can be any task full name or prefix.\n"
+          + "When NAME is given, only tasks with name starts with the prefix will run.\n"
+          + "For example, specifying NAME \"master\" or \"ma\" will run both tasks named "
+          + "\"master.rpc.port.available\" and \"master.web.port.available\" but not "
+          + "\"worker.rpc.port.available\".\n"
+          + "If NAME is not given, all tasks for the given TARGET will run.\n\n"
+          + "OPTIONS can be a list of command line options. Each option has the"
+          + " format \"-<optionName> [optionValue]\"\n";
+
   private static final String ALLUXIO_MASTER_CLASS = "alluxio.master.AlluxioMaster";
   private static final String ALLUXIO_WORKER_CLASS = "alluxio.worker.AlluxioWorker";
   private static final String ALLUXIO_PROXY_CLASS = "alluxio.proxy.AlluxioProxy";
+  private static final Options OPTIONS = new Options();
 
   private final Map<ValidationTask, String> mTasks = new HashMap<>();
   private final Map<String, String> mTaskDescriptions = new HashMap<>();
@@ -65,7 +100,7 @@ public final class EnvValidationTool {
    * @param path the UFS path
    * @param conf the UFS configurtions
    * */
-  public EnvValidationTool(String path, AlluxioConfiguration conf) {
+  public ValidateEnv(String path, AlluxioConfiguration conf) {
     mPath = path;
     mConf = conf;
 
@@ -167,10 +202,14 @@ public final class EnvValidationTool {
   }
 
   private ValidationTask registerTask(String name, String description,
-                   ValidationTask task, List<ValidationTask> tasks) {
+                                      AbstractValidationTask task, List<ValidationTask> tasks) {
     mTasks.put(task, name);
     mTaskDescriptions.put(name, description);
     tasks.add(task);
+    List<Option> optList = task.getOptionList();
+    synchronized (ValidateEnv.class) {
+      optList.forEach(opt -> OPTIONS.addOption(opt));
+    }
     return task;
   }
 
@@ -210,11 +249,11 @@ public final class EnvValidationTool {
     String argStr = String.join(" ", cmd.getArgs());
     String homeDir = mConf.get(PropertyKey.HOME);
     String remoteCommand = String.format(
-        "%s/bin/alluxio validateEnv %s %s %s",
-        homeDir, target, name == null ? "" : name, argStr);
+            "%s/bin/alluxio validateEnv %s %s %s",
+            homeDir, target, name == null ? "" : name, argStr);
     String localCommand = String.format(
-        "ssh -o ConnectTimeout=5 -o StrictHostKeyChecking=no -tt %s \"bash %s\"",
-        node, remoteCommand);
+            "ssh -o ConnectTimeout=5 -o StrictHostKeyChecking=no -tt %s \"bash %s\"",
+            node, remoteCommand);
     String[] command = {"bash", "-c", localCommand};
     try {
       ProcessBuilder builder = new ProcessBuilder(command);
@@ -230,16 +269,9 @@ public final class EnvValidationTool {
     }
   }
 
-  /**
-   * Runs the validation tasks locally.
-   *
-   * @param target target task set
-   * @param name task name prefix
-   * @param cmd the command line
-   * @return whether the validations passed
-   * */
-  public boolean validateLocal(String target, String name, CommandLine cmd)
-      throws InterruptedException {
+  // runs validation tasks in local environment
+  private boolean validateLocal(String target, String name, CommandLine cmd)
+          throws InterruptedException {
     int validationCount = 0;
     Map<ValidationUtils.State, Integer> results = new HashMap<>();
     Map<String, String> optionsMap = new HashMap<>();
@@ -297,25 +329,11 @@ public final class EnvValidationTool {
     return true;
   }
 
-  /**
-   * Runs the worker validation tasks.
-   *
-   * @param name task name prefix
-   * @param cmd the command line
-   * @return whether the validations passed
-   * */
-  public boolean validateWorkers(String name, CommandLine cmd) throws InterruptedException {
+  private boolean validateWorkers(String name, CommandLine cmd) throws InterruptedException {
     return validateRemote(ConfigurationUtils.getWorkerHostnames(mConf), "worker", name, cmd);
   }
 
-  /**
-   * Runs the master validation tasks.
-   *
-   * @param name target task prefix
-   * @param cmd the command line
-   * @return whether the validations passed
-   * */
-  public boolean validateMasters(String name, CommandLine cmd) throws InterruptedException {
+  private boolean validateMasters(String name, CommandLine cmd) throws InterruptedException {
     return validateRemote(ConfigurationUtils.getMasterHostnames(mConf), "master", name, cmd);
   }
 
@@ -329,48 +347,130 @@ public final class EnvValidationTool {
     System.out.println();
   }
 
-  /**
-   * Prints all the tasks.
-   * */
-  public void printTasks() {
+  private void printTasks() {
     printTasks("master");
     printTasks("worker");
     printTasks("cluster");
   }
 
-  /**
-   * Aggregates the options from each validation tasks.
-   * Each {@link ValidationTask} defines the {@link Options} in a OPTIONS field
-   *
-   * @return all options
-   * */
-  public static Options getOptions() {
-    Options options = new Options();
-    Reflections reflections = new Reflections(ValidationTask.class.getPackage().getName());
-    for (Class<? extends ValidationTask> cls : reflections.getSubTypesOf(ValidationTask.class)) {
-      boolean hasOptions = Arrays.stream(cls.getFields())
-              .anyMatch(f -> f.getName().equals("OPTIONS"));
-      if (!hasOptions) {
-        continue;
-      }
+  private static int runTasks(String target, String name, CommandLine cmd)
+          throws InterruptedException {
+    // Validate against root path
+    AlluxioConfiguration conf = InstancedConfiguration.defaults();
+    String rootPath = conf.get(PropertyKey.MASTER_MOUNT_TABLE_ROOT_UFS);
+    ValidateEnv validate = new ValidateEnv(rootPath, conf);
 
-      Field[] fields = cls.getFields();
-      Options taskOptions = null;
-      for (Field f : fields) {
-        if (f.getName().equals("OPTIONS")) {
-          try {
-            taskOptions = (Options) f.get(cls);
-            break;
-          } catch (IllegalAccessException e) {
-            // Cannot access field, skip it
-            break;
-          }
-        }
-      }
-      if (taskOptions != null) {
-        taskOptions.getOptions().forEach(options::addOption);
-      }
+    boolean success;
+    switch (target) {
+      case "local":
+      case "worker":
+      case "master":
+        success = validate.validateLocal(target, name, cmd);
+        break;
+      case "all":
+        success = validate.validateMasters(name, cmd);
+        success = validate.validateWorkers(name, cmd) && success;
+        success = validate.validateLocal("cluster", name, cmd) && success;
+        break;
+      case "workers":
+        success = validate.validateWorkers(name, cmd);
+        break;
+      case "masters":
+        success = validate.validateMasters(name, cmd);
+        break;
+      default:
+        printHelp("Invalid target.");
+        return -2;
     }
-    return options;
+    return success ? 0 : -1;
+  }
+
+  /**
+   * Prints the help message.
+   *
+   * @param message message before standard usage information
+   */
+  public static void printHelp(String message) {
+    System.err.println(message);
+    HelpFormatter formatter = new HelpFormatter();
+    formatter.printHelp(USAGE, OPTIONS, true);
+  }
+
+  /**
+   * Validates environment.
+   *
+   * @param argv list of arguments
+   * @return 0 on success, -1 on validation failures, -2 on invalid arguments
+   */
+  public static int validate(String... argv) throws InterruptedException {
+    if (argv.length < 1) {
+      printHelp("Target not specified.");
+      return -2;
+    }
+    String command = argv[0];
+    String name = null;
+    String[] args;
+    int argsLength = 0;
+    // Find all non-option command line arguments.
+    while (argsLength < argv.length && !argv[argsLength].startsWith("-")) {
+      argsLength++;
+    }
+    if (argsLength > 1) {
+      name = argv[1];
+      args = Arrays.copyOfRange(argv, 2, argv.length);
+    } else {
+      args = Arrays.copyOfRange(argv, 1, argv.length);
+    }
+
+    CommandLine cmd;
+    try {
+      cmd = parseArgsAndOptions(OPTIONS, args);
+    } catch (InvalidArgumentException e) {
+      System.err.format("Invalid argument: %s.%n", e.getMessage());
+      return -1;
+    }
+    if (command.equals("list")) {
+      // Validate against root path
+      AlluxioConfiguration conf = InstancedConfiguration.defaults();
+      String rootPath = conf.get(PropertyKey.MASTER_MOUNT_TABLE_ROOT_UFS);
+      ValidateEnv task = new ValidateEnv(rootPath, conf);
+      task.printTasks();
+      return 0;
+    }
+    return runTasks(command, name, cmd);
+  }
+
+  /**
+   * Validates Alluxio environment.
+   *
+   * @param args the arguments to specify which validation tasks to run
+   */
+  public static void main(String[] args) throws InterruptedException {
+    System.exit(validate(args));
+  }
+
+  /**
+   * Parses the command line arguments and options in {@code args}.
+   *
+   * After successful execution of this method, command line arguments can be
+   * retrieved by invoking {@link CommandLine#getArgs()}, and options can be
+   * retrieved by calling {@link CommandLine#getOptions()}.
+   *
+   * @param args command line arguments to parse
+   * @return {@link CommandLine} object representing the parsing result
+   * @throws InvalidArgumentException if command line contains invalid argument(s)
+   */
+  private static CommandLine parseArgsAndOptions(Options options, String... args)
+          throws InvalidArgumentException {
+    CommandLineParser parser = new DefaultParser();
+    CommandLine cmd;
+
+    try {
+      cmd = parser.parse(options, args);
+    } catch (ParseException e) {
+      throw new InvalidArgumentException(
+              "Failed to parse args for validateEnv", e);
+    }
+    return cmd;
   }
 }
